Convert cart into a context provided value
This is so we can use it in product detail and determine whether or not the cart already has the max number of items. In doing so, we had to move the Strict Mode to inside of the router to prevent it from double firing the dispatch from useReducer (which would double-add items added to cart).