Re-implement StrictMode in index
StrictMode's double firing was no mistake. It's to help you identify any functions that should be pure, but aren't. Since a reducer must be a pure function, the double-fire showed it was impure since it mutated the state (the cart).

Since the cart is an object of objects, and the original logic only *shallow copied* the result, it means that each object in the new copy was still referenced the old state. It therefore incremented the value of the old state, inadvertently mutating it, and then returning that value!

React's strict mode double fires and runs the function again with that state. So we get a double count because the state was mutated.

RUN 1
State 1 -> run 1 -> return state
discard that result! we want to test for purity.

RUN 2
State 2 (the state should the same, React strict mode says let's run again to be safe) -> run 2 -> return state
actually do it this time.

SOLUTION:

Use Lodash to deep clone the item. The updated state is therefore divorced from its previous iteration through a separate copy of itself.