Handle 404s for Invalid Products, and Case when Product URL accessed directly
There's more than one way to access the product page, which means we can't take advantage of the state passed through the Link object in the category component. That means we have to fetch the data again. We use useEffect to execute the newly added async function fetchData.

The fact that we have to use async operations to load data means that we need to keep track of the state of loading, so we added the enum PAGE_STATES, which has three states -- ready, loading, and not found.

- Any page accessed through the parent category page will instantly be ready.
- Any page which is accessed directly through URL instead will be loading initially and then be set to ready after it is finished loading.
- Any page which is accessed directly through URL, but the URL is incorrect will be set to NOT_FOUND.