Change BrowserRouter to HashRouter
Just learnt that BrowserRouter is meant for applications with a back-end, otherwise it needlessly sends requests to the server. Instead, with a HashRouter, the URL gets a hash appended to denote a front-end route -- the browser (client) should process and interpret this instead.

For hosting on Github Pages, you need to use a HashRouter otherwise the pathing gets all messed up (sends request to root of the user's github pages site instead of the subfolder directory, resulting in 404 on Github's Page).